uradjene api rute

+++ database/migrations/2023_03_19_073553_create_items_table.php
@@ -11,11 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('items', function (Blueprint $table)
-        {
+        Schema::create('items', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('price');
+            $table->decimal('price', 9, 2);
             $table->foreignId('restaurant_id');
             $table->timestamps();
